[fix/#128] - middleware 중 특정 경로에 대한 라우트 처리로 불필요한 서버요청 줄이도록 수정
